Revert 'Ensure signing only happens when using `*_sign` policies'
As discussed in the review, adding new sync policies is the wrong place for
ensuring that signing happens only when necessary.  For the moment, we are
going to assume that if a user has set a signing service and fingerprint
for a repository, then they want it re-signed on sync.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>